# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Shri BalaJi Dham HMS API",
        "description": "Hotel Management System REST API for Shri BalaJi Dham Hotel, Mathura. Provides endpoints for room management, bookings, guest management, financials, and more.",
        "contact": {
            "name": "Shri BalaJi Dham Support",
            "email": "sribalajidhamhotel@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/v1/bookings": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "List bookings with filters and pagination",
                "operationId": "listBookings",
                "parameters": [
                    {
                        "name": "property_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "confirmed",
                                "checked-in",
                                "checked-out",
                                "cancelled",
                                "no-show"
                            ]
                        }
                    },
                    {
                        "name": "payment_status",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "unpaid",
                                "partially-paid",
                                "paid"
                            ]
                        }
                    },
                    {
                        "name": "from_date",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to_date",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "guest_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by booking reference or guest name",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 25
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bookings list with pagination"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Create a new booking",
                "operationId": "storeBooking",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "property_id",
                                    "guest_id",
                                    "checkin_date",
                                    "checkout_date",
                                    "number_of_adults",
                                    "rooms"
                                ],
                                "properties": {
                                    "property_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "guest_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "checkin_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-04-01"
                                    },
                                    "checkout_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-04-03"
                                    },
                                    "number_of_adults": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "number_of_children": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "source": {
                                        "type": "string",
                                        "example": "website"
                                    },
                                    "special_requests": {
                                        "type": "string",
                                        "example": "Early check-in if possible"
                                    },
                                    "rooms": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "room_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "rate_cents": {
                                                    "type": "integer",
                                                    "example": 237600
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Booking created"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "409": {
                        "description": "Room not available"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/bookings/{id}": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Get booking details with rooms, charges, payments",
                "operationId": "showBooking",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Booking details"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/bookings/{id}/checkin": {
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Check-in a guest",
                "operationId": "checkinBooking",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Checked in successfully"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "409": {
                        "description": "Invalid status for check-in"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/bookings/{id}/checkout": {
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Check-out a guest",
                "operationId": "checkoutBooking",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Checked out successfully"
                    },
                    "409": {
                        "description": "Invalid status for check-out"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/bookings/{id}/cancel": {
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Cancel a booking",
                "operationId": "cancelBooking",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "reason": {
                                        "type": "string",
                                        "example": "Guest requested cancellation"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Booking cancelled"
                    },
                    "409": {
                        "description": "Cannot cancel"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/bookings/today": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Get today's check-ins and check-outs",
                "operationId": "todayBookings",
                "parameters": [
                    {
                        "name": "property_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Today's bookings summary"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "description": "Paste the token from /auth/login response (without \"Bearer\" prefix)"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}